Extract content to seperate files, excellent content
