+++ frontend/src/App.tsx
@@ -1,52 +1,19 @@
 import './App.css'
 import * as React from 'react';
-import Site from './types/Site.types';
 import AuthContext from './types/AuthContext.types';
 import TokenPayload from './types/TokenPayload.types';
 import { GoogleLogin, CredentialResponse } from '@react-oauth/google';
 import { jwtDecode } from 'jwt-decode';
-import ConfigPanel from './components/ConfigPanel';
 import Navigation from './components/Navigation';
-import QuizListItem from './components/lists/QuizListItem';
-import ListListItem from './components/lists/SiteListItem';
-import SiteList from './components/lists/SiteList';
-import { useEffect } from 'react';
-import { mockListSites, mockQuizSites } from './mock/mockData';
+import QuizSitePage from './pages/QuizSitePage';
+import ListSitePage from './pages/ListSitePage';
 
 function App() {
-  const [hideLowQuality, setHideLowQuality] = React.useState(true);
-  const [listSites, setListSites] = React.useState<Site[]>([]);
-  const [quizSites, setQuizSites] = React.useState<Site[]>([]);
   const [authContext, setAuthContext] = React.useState<AuthContext>({
     state: import.meta.env.VITE_SKIP_LOGIN === 'true' ? 'LOGGED_IN' : 'NOT_LOGGED_IN',
     email: null,
     token: null
   });
-
-  useEffect(() => {
-    const useMockData = import.meta.env.VITE_USE_MOCK_DATA === 'true';
-    if (useMockData) {
-      setListSites(mockListSites);
-    } else {
-      fetch("/api/list/site")
-        .then(response => response.json())
-        .then((data) => setListSites(data))
-        .catch((error) => console.error(error));
-    }
-    if (useMockData) {
-      setQuizSites(mockQuizSites);
-    } else {
-      fetch("/api/list/quiz")
-        .then(response => response.json())
-        .then((data) => setQuizSites(data))
-        .catch((error) => console.error(error));
-    }
-  }, []);
-
-
-  const filterLowQuality = (site: Site) => {
-    return !(site.lowQuality && hideLowQuality);
-  };
 
   const handleSuccessLogin = (credentialResponse: CredentialResponse) => {
     if (credentialResponse.credential) {
@@ -72,33 +39,6 @@ function App() {
     <div>BAD</div>
   )
 
-  const listContent = () => (
-    <>
-      <ConfigPanel
-        lowQualityListsHidden={hideLowQuality}
-        onLowQualityCheckboxChange={setHideLowQuality} />
-      <SiteList>
-        {listSites.filter(filterLowQuality).map(site => (
-          <ListListItem
-            key={site.name}
-            listSite={site}
-          />
-        ))}
-      </SiteList>
-    </>
-  );
-
-  const quizContent = () => (
-    <SiteList>
-      {quizSites.map(site => (
-        <QuizListItem
-          key={site.name}
-          quizSite={site}
-        />
-      ))}
-    </SiteList>
-  );
-
   const loginButton = () => (
     <GoogleLogin
       onSuccess={handleSuccessLogin}
@@ -114,12 +54,12 @@ function App() {
         {
           label: 'Lists',
           index: '1',
-          content: listContent()
+          content: <ListSitePage />
         },
         {
           label: 'Quiz',
           index: '2',
-          content: quizContent()
+          content: <QuizSitePage />
         }
       ]}
     />
